Add Fishing Shop items for GMS2 Region
I've added the Fisherman's Shop items for GMS2 Region as well as the Shop was empty before.
It might not 100% be perfect and resemble to final state of the game but it was the closest I
could get it with the available resources that could be found.
Also the Pesca NPC at Rainbow Fishing Hole was updated

Refs:
- https://youtu.be/UfhNeqAragU?si=Pu-JHOxuxlxKFqLK&t=283
- https://youtu.be/PM5JSEGCo2E?si=dfRQw8XpqLA7k48_&t=463